requirements.txt
fastapi
uvicorn
jinja2
python-multipart
openai
faiss-cpu
langchain
langchain-community
langchain-huggingface
huggingface-hub
sentence-transformers
python-dotenv
requests
pydantic
numpy
tqdm
transformers
PyMuPDF
python-pptx